finished resources pages

+++ app/views/resource.py
@@ -19,6 +19,7 @@ from flask_login import current_user
 from flask_login import login_required
 
 from app.forms import NewResourceForm 
+from app.forms import EditResourceForm 
 
 import os
 
@@ -32,6 +33,21 @@ def new_resource():
         'resource/new_resource.html',
         title='New Resource',
         form=NewResourceForm()
+    )
+
+@resource.route('/edit_resource/<id>', methods=['GET'])
+@login_required
+def edit_resource(id):
+    resource = Resource.query.filter_by(id=id).first()
+    if not resource:
+        flash('URL does not exist')
+        return redirect(url_for('resource.resource_home'))
+
+    return render_template(
+        'resource/edit_resource.html',
+        resource=resource,
+        title='Edit Resource',
+        form=EditResourceForm()
     )
 
 @resource.route('/resource/home', methods=['GET'])
@@ -101,7 +117,6 @@ def download_resource(id):
         attachment_filename=f"{resource.title}.pdf")
 
 # ***************-CRUD API OPERATIONS-*************** #
-# create
 @resource.route('/resource', methods=['POST'])
 @login_required
 def create_resource():
@@ -121,7 +136,6 @@ def create_resource():
         A status code with a message
     '''
     # Make sure form data exists in request
-    message = ''
     if not request.form:
         message = 'No form data present in request'
         return message, status.HTTP_400_BAD_REQUEST
@@ -132,6 +146,9 @@ def create_resource():
     required_params = ['category', 'title']
     for param in params.keys():
         if param in form_data:
+            if form_data[param] == '' and param in required_params:
+                message = f'Invalid request. No {param} present'
+                return message, status.HTTP_400_BAD_REQUEST 
             params[param] = form_data[param]
         else:
             if param in required_params:
@@ -166,9 +183,9 @@ def create_resource():
     )
     db.session.add(resource)
     db.session.commit()
+    message = 'success'
     return message, status.HTTP_200_OK
 
-# read
 @resource.route('/resource', methods=['GET'])
 @login_required
 def read_resource():
@@ -177,6 +194,8 @@ def read_resource():
     Query String Params:
         - category (optional): the category of resources you want to get
         - resource_type (optional): file, video, or article
+        - recent (optional): will return the 4 most recent resources if 
+            resource=true is set
         
     Returns:
         An array of resources based on the query string params. Will return all resources
@@ -186,6 +205,7 @@ def read_resource():
     # Gather data from query string params
     category = request.args.get('category')
     resource_type = request.args.get('resource_type')
+    recent = request.args.get('recent')
 
     # this will be where query is held
     query = 'Resource.query'
@@ -199,7 +219,7 @@ def read_resource():
         query += f".filter(Resource.video_link != '')"
     if resource_type in ['file', 'pdf', 'PDF', 'File']:
         resource_type = 'file'
-        query += f".filter(Resource.file_path != None)"
+        query += f".filter(Resource.file_path != '')"
     
     if category in ['Pitching', 'pitching', 'pitchers', 'Pitchers']:
         category = 'Pitching'
@@ -207,32 +227,34 @@ def read_resource():
         category = 'Hitting'
     if category in ['Defense', 'defense', 'fielding', 'fielders', 'Fielding']:
         category = 'Defense'
-    if category in ['Mental Game', 'Mental', 'mental game', 'mental']:
+    if category in ['Mental Game', 'Mental', 'mental game', 'mental']: 
         category = 'Mental Game'
     if category in ['Miscellaneous', 'miscellaneous', 'Other', 'other']:
         category = 'Miscellaneous'
 
     if category:
         query += f".filter_by(category = '{category}')"
-
+    
     query += '.all()'
 
     # query database and return an array of resources
     resources = eval(query)
     return_resources = []
+    index = 0
     for resource in resources:
-        return_resources.append(resource.to_dict())
+        if recent == 'true':
+            if index >= len(resources) - 4:
+                return_resources.append(resource.to_dict())
+        else:
+            return_resources.append(resource.to_dict())
+        index += 1
 
     return jsonify(return_resources), status.HTTP_200_OK
     
-# update
 @resource.route('/resource/<id>', methods=['PATCH'])
 @login_required
 def update_resource(id):
     '''API Route to update a resource
-    The update will take the exact form of the request, so if something doesn't need to
-    be updated, it still needs to be passed as form data.
-
     Only thing that could be added here is url validation for the articles or videos, 
         but that is done on the front end (form) side
 
@@ -245,11 +267,10 @@ def update_resource(id):
             video_link: a valid youtube url to a video
             file: a pdf file 
     Returns:
-        HTML based on the request and whether it was successfull or not
+        Status code and a message based on request
     '''
 
     # Check to make sure resource exists
-    message = ''
     resource = Resource.query.filter_by(id=id).first()
     if not resource:
         message = 'Invalid request. ID passed does not match any resource'
@@ -266,24 +287,36 @@ def update_resource(id):
     required_params = ['category', 'title']
     for param in params.keys():
         if param in form_data:
+            if form_data[param] == '' and param in required_params:
+                message = f'Invalid request. No {param} present'
+                return message, status.HTTP_400_BAD_REQUEST
             params[param] = form_data[param]
         else:
             if param in required_params:
                 message = f'Invalid request. No {param} present'
-                return message, status.HTTP_400_BAD_REQUEST          
+                return message, status.HTTP_400_BAD_REQUEST
 
-    # Check for file in request
-    if file_exists(resource.file_path):
-        os.remove(resource.file_path)
-    file_path = save_file_to_file_system(request, 'file', params['title'])
-    if file_path == 'Error':
-        message = 'There was an error while saving the file'
-        return message, status.HTTP_500_INTERNAL_SERVER_ERROR
-    
-    # Only one of three params must be present: article_line, video_link, file_data
-    if not exactly_one_param_present(params['article_link'], params['video_link'], file_path):
-        message = f'Invalid request. Need exactly one of: article link, video link, file'
-        return message, status.HTTP_400_BAD_REQUEST
+    # Only check article/video/file params if sent. If not, just use the old one
+    if params['article_link'] or params['video_link'] or 'file' in request.files:
+        # Check for file in request
+        file_path = save_file_to_file_system(request, 'file', params['title'])
+        if file_path == 'Error':
+            message = 'There was an error while saving the file'
+            return message, status.HTTP_500_INTERNAL_SERVER_ERROR
+
+        # Only one of three params must be present: article_line, video_link, file_data
+        if not exactly_one_param_present(params['article_link'], params['video_link'], file_path):
+            message = f'Invalid request. Need exactly one of: article link, video link, file'
+            return message, status.HTTP_400_BAD_REQUEST
+
+        # delete the old file if it exists
+        if file_exists(resource.file_path):
+            os.remove(resource.file_path)
+
+    else:
+        params['article_link'] = resource.article_link
+        params['video_link'] = resource.video_link
+        file_path = resource.file_path
 
     # Update the resource based on everything passed through
     resource.category = params['category']
@@ -294,9 +327,9 @@ def update_resource(id):
     resource.file_path = file_path
 
     db.session.commit()
+    message = 'success'
     return message, status.HTTP_200_OK
 
-# delete
 @resource.route('/resource/<id>', methods=['DELETE'])
 @login_required
 def delete_resource(id):
@@ -305,9 +338,9 @@ def delete_resource(id):
     Requst Args:
         The id in the url such as /resource/4 to delete the resource with id=4
     Returns:
-        HTML based on the request and whether it was successfull or not
+        Status code and a message based on request
     '''
-    message = ''
+
     # check to make sure the resource exists
     resource = Resource.query.filter_by(id=id).first()
     if not resource:
@@ -320,10 +353,14 @@ def delete_resource(id):
 
     db.session.delete(resource)
     db.session.commit()
+    message = 'success'
     return message, status.HTTP_200_OK
 
 # ***************-UTIL FUNCTIONS-*************** #
 def exactly_one_param_present(one, two, three):
+    '''
+    Checks the three params and returns true if exactly one of them exist
+    '''
     number_present = 0
     if one:
         number_present += 1
@@ -334,6 +371,18 @@ def exactly_one_param_present(one, two, three):
     return number_present == 1
 
 def save_file_to_file_system(request, file_name_in_request, file_name):
+    '''
+    Checks a http request for a file passed along and saves it to the file system
+    Args:
+        - request: the Flask http request
+        - file_name_in_request: the file name passed along in the request to save
+        - file_name: the name to give the file that is being saved
+    Returns:
+        - If successful save, the relative file path when in the 'static' directory
+            making it easy to render on the client side for download
+        - If it couldn't save, the string 'Error' is returned
+        - If the file was not in the request, None is returned
+    '''
     if file_name_in_request in request.files:
         if not request.files[file_name_in_request].filename == '':
             file_loc = os.path.join(
@@ -366,6 +415,9 @@ def save_file_to_file_system(request, file_name_in_request, file_name):
     return None
 
 def file_exists(file_path):
+    '''
+    Returns True if the file exists for the path given, False otherwise
+    '''
     try:
         if not file_path:
             return False
